chore: Add noop typescript worklets module
