+++ packages/react-native-reanimated/src/ReanimatedModule/NativeReanimated.ts
@@ -1,11 +1,11 @@
 'use strict';
 import type {
-  ShadowNodeWrapper,
   Value3D,
   ValueRotation,
   ShareableRef,
   LayoutAnimationBatchItem,
   IReanimatedModule,
+  IWorkletsModule,
 } from '../commonTypes';
 import { checkCppVersion } from '../platform-specific/checkCppVersion';
 import { jsVersion } from '../platform-specific/jsVersion';
@@ -16,58 +16,11 @@ import type React from 'react';
 import { getShadowNodeWrapperFromRef } from '../fabricUtils';
 import { ReanimatedTurboModule } from '../specs';
 import { ReanimatedError } from '../errors';
+import { WorkletsModule } from '../worklets';
+import type { ReanimatedModuleProxy } from './reanimatedModuleProxy';
 
-export function createNativeReanimatedModule() {
+export function createNativeReanimatedModule(): IReanimatedModule {
   return new NativeReanimatedModule();
-}
-
-// this is the type of `__reanimatedModuleProxy` which is injected using JSI
-export interface ReanimatedModuleProxy {
-  makeShareableClone<T>(
-    value: T,
-    shouldPersistRemote: boolean,
-    nativeStateSource?: object
-  ): ShareableRef<T>;
-  scheduleOnUI<T>(shareable: ShareableRef<T>): void;
-  executeOnUIRuntimeSync<T, R>(shareable: ShareableRef<T>): R;
-  createWorkletRuntime(
-    name: string,
-    initializer: ShareableRef<() => void>
-  ): WorkletRuntime;
-  scheduleOnRuntime<T>(
-    workletRuntime: WorkletRuntime,
-    worklet: ShareableRef<T>
-  ): void;
-  registerEventHandler<T>(
-    eventHandler: ShareableRef<T>,
-    eventName: string,
-    emitterReactTag: number
-  ): number;
-  unregisterEventHandler(id: number): void;
-  getViewProp<T>(
-    viewTagOrShadowNodeWrapper: number | ShadowNodeWrapper,
-    propName: string,
-    callback?: (result: T) => void
-  ): Promise<T>;
-  enableLayoutAnimations(flag: boolean): void;
-  registerSensor(
-    sensorType: number,
-    interval: number,
-    iosReferenceFrame: number,
-    handler: ShareableRef<(data: Value3D | ValueRotation) => void>
-  ): number;
-  unregisterSensor(sensorId: number): void;
-  configureProps(uiProps: string[], nativeProps: string[]): void;
-  subscribeForKeyboardEvents(
-    handler: ShareableRef<number>,
-    isStatusBarTranslucent: boolean,
-    isNavigationBarTranslucent: boolean
-  ): number;
-  unsubscribeFromKeyboardEvents(listenerId: number): void;
-  configureLayoutAnimationBatch(
-    layoutAnimationsBatch: LayoutAnimationBatchItem[]
-  ): void;
-  setShouldAnimateExitingForTag(viewTag: number, shouldAnimate: boolean): void;
 }
 
 function assertSingleReanimatedInstance() {
@@ -83,9 +36,15 @@ See \`https://docs.swmansion.com/react-native-reanimated/docs/guides/troubleshoo
 }
 
 class NativeReanimatedModule implements IReanimatedModule {
+  /**
+   * We keep the instance of `WorkletsModule` here to keep correct coupling of
+   * the modules and initialization order.
+   */
+  #workletsModule: IWorkletsModule;
   #reanimatedModuleProxy: ReanimatedModuleProxy;
 
   constructor() {
+    this.#workletsModule = WorkletsModule;
     // These checks have to split since version checking depend on the execution order
     if (__DEV__) {
       assertSingleReanimatedInstance();
