+++ packages/react-native-reanimated/src/ReanimatedModule/index.ts
@@ -1,8 +1,4 @@
 'use strict';
-import { createJSReanimatedModule } from './js-reanimated';
-import { shouldBeUseWeb } from '../PlatformChecker';
-import { createNativeReanimatedModule } from './NativeReanimated';
 
-export const ReanimatedModule = shouldBeUseWeb()
-  ? createJSReanimatedModule()
-  : createNativeReanimatedModule();
+export { ReanimatedModule } from './reanimatedModuleInstance';
+export type { ReanimatedModuleProxy } from './reanimatedModuleProxy';
