+++ packages/react-native-reanimated/src/ReanimatedModule/reanimatedModuleInstance.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { createJSReanimatedModule } from './js-reanimated';
+import { shouldBeUseWeb } from '../PlatformChecker';
+import { createNativeReanimatedModule } from './NativeReanimated';
+
+export const ReanimatedModule = shouldBeUseWeb()
+  ? createJSReanimatedModule()
+  : createNativeReanimatedModule();
