+++ packages/react-native-reanimated/src/ReanimatedModule/reanimatedModuleInstance.web.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+import { createJSReanimatedModule } from './js-reanimated';
+
+export const ReanimatedModule = createJSReanimatedModule();
