+++ packages/react-native-reanimated/src/privateGlobals.d.ts
@@ -15,7 +15,7 @@ import type {
 } from './commonTypes';
 import type { AnimatedStyle } from './helperTypes';
 import type { FrameCallbackRegistryUI } from './frameCallback/FrameCallbackRegistryUI';
-import type { ReanimatedModuleProxy } from './ReanimatedModule/NativeReanimated';
+import type { ReanimatedModuleProxy } from './ReanimatedModule';
 import type { SensorContainer } from './SensorContainer';
 import type { LayoutAnimationsManager } from './layoutReanimation/animationsManager';
 import type { ProgressTransitionRegister } from './layoutReanimation/sharedTransitions';
