+++ packages/react-native-reanimated/src/worklets/WorkletsModule/JSWorklets.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import type { IWorkletsModule } from '../../commonTypes';
+
+export function createJSWorkletsModule(): IWorkletsModule {
+  return new JSWorklets();
+}
+
+class JSWorklets implements IWorkletsModule {}
