+++ packages/react-native-reanimated/src/worklets/WorkletsModule/NativeWorklets.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import type { IWorkletsModule } from '../../commonTypes';
+
+export function createNativeWorkletsModule(): IWorkletsModule {
+  return new NativeWorklets();
+}
+
+class NativeWorklets {}
