+++ packages/react-native-reanimated/src/worklets/WorkletsModule/workletsModuleInstance.ts
@@ -1,0 +1,9 @@
+'use strict';
+
+import { createNativeWorkletsModule } from './NativeWorklets';
+import { shouldBeUseWeb } from '../../PlatformChecker';
+import { createJSWorkletsModule } from './JSWorklets';
+
+export const WorkletsModule = shouldBeUseWeb()
+  ? createJSWorkletsModule()
+  : createNativeWorkletsModule();
