+++ packages/react-native-reanimated/src/worklets/WorkletsModule/workletsModuleInstance.web.ts
@@ -1,0 +1,5 @@
+'use strict';
+
+import { createJSWorkletsModule } from './JSWorklets';
+
+export const WorkletsModule = createJSWorkletsModule();
