+++ packages/react-native-reanimated/src/worklets/index.ts
@@ -1,0 +1,3 @@
+'use strict';
+
+export { WorkletsModule } from './WorkletsModule';
